all the feedback are done

+++ f-manage/src/app/pages/users/users.ts
@@ -8,13 +8,17 @@ import { AddEditUser } from './add-edit-user/add-edit-user';
 import { ToastComponent } from '../../shared/components/toast-notification/toast-notification';
 import { ToastService } from '../../shared/services/toast.service';
 import { Layout } from '../../shared/components/layout/layout';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../shared/components/confirm-dialog/confirm-dialog';
+import { TooltipDirective } from '../../shared/components/tooltip/tooltip';
 
 @Component({
   selector: 'app-users',
   standalone: true,
   imports: [CommonModule, FormsModule, CustomTableComponent , AddEditUser ,
     Layout,
-    ToastComponent],
+    ToastComponent,
+    TooltipDirective],
   templateUrl: "./users.html",
   styleUrls: ["./users.scss"],
 })
@@ -69,7 +73,11 @@ export class Users implements OnInit {
 
   totalItems = computed(() => this.filteredUsers().length);
 
-  constructor(private dbService: IndexedDBService, private toastService: ToastService) {}
+  constructor(
+    private dbService: IndexedDBService, 
+    private toastService: ToastService,
+    private dialog: MatDialog
+  ) {}
 
   ngOnInit() {
     this.loadUsers();
@@ -97,8 +105,13 @@ export class Users implements OnInit {
       this.selectedUserId = event.item.id ?? null;
       this.showUserModal = true;
     } else if (event.action === 'delete') {
-      if (confirm('Are you sure you want to delete this user?')) {
-        if (event.item.id) {
+      const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+        width: '400px',
+        data: { message: 'Are you sure you want to delete this user?' }
+      });
+
+      dialogRef.afterClosed().subscribe(result => {
+        if (result && event.item.id) {
           this.dbService.deleteUser(event.item.id).subscribe({
             next: () => {
               this.loadUsers();
@@ -106,10 +119,11 @@ export class Users implements OnInit {
             },
             error: (error: Error) => {
               console.error('Error deleting user:', error);
+              this.toastService.showError('Failed to delete user. Please try again.');
             }
           });
         }
-      }
+      });
     }
   }
